Move post body shortener to utils
